Fix text insertion with multiple cursors at same position.
We can have 2 cursors at the same positions when there are two selections
that start/end at the same position. In such a case cursor positions are same
but selections are different. Thus text insertion needs to be handled
differntly. I have tried to match the behaviour of sublime/vscode for this.

BUG: 492869