Minimap now follows the theme also for search matches
The search match is now the same color as normal text

There is no color in syntax-highlighting for the search match marker and the search match highlight background is not very visible in the minimap. That's why we use the foreground color to make them stand out.

Light

![image.png](/uploads/f043f4e3b7b1313177139874402b776d/image.png){width=459 height=248}

Dark

![image.png](/uploads/7821d5fecfbe8cffafeedc7978cdd24b/image.png){width=459 height=249}

(which shows a bug with the arrow up/down button drawing in the minimap, but that is a different issue)