Remove MovingRange caching in TextBlock
MovingRange caching per line is very expensive and very complex with
lots of gotchas and has very little usefulness in practice.

Remove all of it, we can just use the cursors to find the ranges in
line. As a result the code is a lot simpler to understand and imo a lot
safer too since there is no more "fixing the block from the inside the text
range" and always worrying about whether the caches are right.

This gives a big performance boost to the moving stuff. It is no longer
very slow to have a lot moving ranges around.