buffer: Remove m_invalidCursors
The field serves no real purpose, but it adds extra complexity
and slows things down for every call to TextCursor::setPosition.

The worst that can happen by removing this is memory leaks but
that is easily fixable on the client side. In kate, for e.g.,
there are no leaks at all.