Dont take ownership of the MovingRange/MovingCursor
Let the user decide what to do with the range or cursor and when to delete
it. This simplifies the Moving API and ownership which is currently
confusing as both sides(user, buffer) own and try to delete them. It also
allows simplification and optimization of internal moving range code.