Fix crash if feedback or dyn attr is cleared before deletion
We only notify about deleted range in the dtor if a range has feedback
or a dynamic attribute. That will not work if someone clears the
feedback or attribute before deletion.

To safeguard against this, notify about this change if the feedback/
attribute gets unset.

Note that the parameter is called "deletedRange" which isn't really
accurate anymore and needs to be renamed but I am not really sure
what to name it.

Ideally the view or anyone else wouldn't store non-owning range pointers
like this. I don't have any good ideas about a safer "handle" that also
performs well atm so leaving it as is.

BUG: 495925