Fix scrolling to search results in wrapped long lines
When dynamic word wrap is enabled, if the search result is located in a very long line and part of the line is currently displayed but the search result is further down, the search result cannot be navigated to after searching.

This commit modifies the return value of KateLayoutCache::displayViewLine to ensure that KateViewInternal::makeVisible can correctly scroll to search results in long lines.

BUG: 448765