Add Text to Speech actions to context menu
This merge request adds the text to speech actions from the Tools menu to the editor's context menu for more-convenient access. The original Tools menu is left untouched, since I saw no reason to remove it.

I decided to add the entire "Text to Speech" submenu to the context menu, but moved the "Say current selection or document" action to the root of the context menu (at the very bottom). This is because, when using desktop zoom, sub-menus are generally a lot more annoying to navigate when panning the zoom area around - and I want the "Say" action to be as easy-to-access as possible, but don't want to clutter the menu with the rest of the TTS actions (since I'd imagine they'd be used less often).

My intention with this MR is to make it so low-vision users like myself can use both Kate and KWrite to read text files, without using a full screen reader, since in my case my vision is well-enough that I can navigate a computer normally (but with heavy zoom), but bad enough that reading large bodies of text is faster with TTS.