Fix issues with fields not being properly initialized
Template field defaults are now calculated and collected. Only
then are they applied to the document.

This fixes issues with mirror fields not always being properly
initialized. It also aligns the actual behavior with the documented
behavior.

This reverts a change that I introduced earlier: it is not possible
anymore to reference fields in default values before they have been
defined. This is more predictable and aligns better with expected
behavior when writing templates. (I.e. values are no longer "hoisted"
like in JS.)