Change tab behavior to select an untouched field before jumping forward
Example: cursor is at "|", pressing tab will now select ${bar} and only
the next press will jump to ${foo}.

    ${foo} |${bar}

Old behavior: pressing tab jumps to ${foo} because the cursor already
touches ${bar}.

The new behavior is more in line with expectations when manually setting
the cursor to a field. It also fixes issues with tabbing through adjacent
fields: in this example, ${bar} would always be skipped. This works now.

    ${baz} ${foo=''}${bar}