Draw selection background ourselves
We were using the "selectionFormats" parameter of QTextLayout::draw()
to draw selections and supplied background + foreground formats for
drawing selections. This leads to extremely poor performance when
there are a lot of "selection formats" e.g., a large list of numbers
on a single line.

To fix this, draw the selection background ourselves and handle the
foreground formats during text layouting phase.

BUG: 502233