Fix selection ignored when layouting the line
We now require that all formats are there when laying out the line.
When text is selected, it can have a different foreground so if we
ignore selection we end up having wrong foreground.

This broke because previously we didn't consider selections during
line layouting. Selections were rendered separately. But now we render
selection ourselves and require that all text coloring etc is handled
during text layouting.