multicursor: Fix performance issues when deleting lot of lines
We are deleting lines in ascending order which can be really bad
for performance if there are a lot of lines. When a line is deleted
it is unwrapped, and the lines below need to be adjusted in for e.g.,
the KateLineLayoutMap. Doing it in descending order is instant because
we reduce the number of lines after the line is deleted.