Optionally return invalid cursor for coordinates outside text
If an identifier is located at the very end of a line, hovering over the
empty space to the right of it while holding the Ctrl key down turns the
mouse pointer into the pointing hand cursor. Clicking on this empty
space while holding Ctrl navigates to the identifier. This is the
unintended behavior of both the LSP Client plugin and KDevelop's
built-in Code Browser plugin. The Open Link plugin behaves similarly,
but only after hovering over a link at the very end of a line, then
moving the mouse pointer to the right of the link.

KDevelop's Code Browser plugin also shows the pointing hand cursor and
navigates on Ctrl+click when the mouse pointer is to the left (up to and
including the mark border) or the right (the minimap) of the text view
if there is an identifier at the very beginning or at the very end of
the line respectively.

5563b89bb1817dc17497cedb6c869dc405900f36 had fixed the bug on the right
side of an identifier. Then 1858d6fafb580c3fc5a53cdd0bf02a227fefd667
unintentionally broke the fix while addressing the following bug:
    Currently if you try to convert the cursor to coordinates and then
    convert these coordinates back to a cursor again, it produces an
    invalid cursor (-1, -1) when the original cursor is located at the
    end of line

Replacing > with == in KateViewInternal::coordinatesToCursor() again
restores the better Ctrl+click behavior but reintroduces the end-of-line
bug. QTextLine::xToCursor(), to which KateRenderer::xToCursor()
delegates, does not allow to distinguish positions "at the end of line"
and "beyond the end of line". Remove the old (now broken and useless)
fix and replace it with a new one described below.

The horizontal boundaries of text on a line can be obtained via
QTextLine::naturalTextRect() or [0, QTextLine::naturalTextWidth()]. So
return an invalid cursor from KateViewInternal::coordinatesToCursor() if
coord.x() falls out of that range. The two alternative ranges were
always equal in my experiments. Opt for the more efficient
naturalTextWidth() range check that appears to work well. If it does not
work for some use case, the slower naturalTextRect() can be checked
instead.

Turns out that most uses of KateViewInternal::coordinatesToCursor() in
KTextEditor rely on its current return value for coordinates of an empty
space. That is, applying the fix unconditionally breaks functionality
that depends on mouse tracking, in particular inline notes. So make the
new behavior optional and opt out of it everywhere in KTextEditor. I am
not sure whether the old or the new behavior is better in
KateViewAccessible::offsetAtPoint(), so keep the old one to be on the
safe side.

The new behavior is beneficial for the single use of
KTextEditor::View::coordinatesToCursor() in KDevelop (the Code Browser),
as well as for the use in the LSP Client plugin. The behavior change
makes no difference to the Open Link plugin, because it checks the word
range at the returned cursor, which is empty at the end of a line.

However, this function is used in the Compiler Explorer plugin to
highlight the assembly code linked to the hovered-over source code line.
The old behavior is probably more useful there. The new behavior also
shows and hides the LSP Client tooltip repeatedly when the mouse pointer
is moved from an identifier at the very end of a line through the empty
space to the right. This regression indicates that the new behavior
cannot be applied in TooltipPrivate::eventFilter() without some
accompanying fixes.

I do not know which behavior works better for the single use of
KTextEditor::View::coordinatesToCursor() in RKWard and Kile each.

So keep the behavior of the existing function unchanged to avoid
regressions. Introduce enum KTextEditor::View::CoordinatesToCursorFlags
and a new overload of KTextEditor::View::coordinatesToCursor() that
takes one more parameter `flags`. This way the uses of this function can
be ported to the new behavior gradually and independently.

frameworks/ktexteditor#25