Use a pool allocator for KateLineLayout
We are creating and destroying these all the time, always allocating on
the heap resulting in waste and poor cache locality. We can instead use
a pool allocator and reuse memory.

This results in a ~2x better performance when we wrap/unwrap a line and
layouts get adjusted in slotEditDone()