Add API to obtain the editor widget from View and vice versa
Multiple applications that use KTextEditor, including Kate, KDevelop,
RKWard and Kile, install an event filter on the editor widget (i.e. the
KateViewInternal child widget) of a KTextEditor::View. Most users obtain
the editor widget of a View efficiently - by calling View::focusProxy().
KDevelop's Code Browser plugin does not rely on this undocumented
implementation detail of KTextEditor and installs an event filter on the
View object itself and on all its descendant widgets recursively, which
might be more reliable, but is certainly less efficient.

Some eventFilter() implementations need to obtain the View from its
editor widget. Many do so by casting the widget's parent() to View*.
KDevelop's Code Browser plugin does not rely on this undocumented
implementation detail of KTextEditor and casts the ancestors of the
widget to View* recursively - until the cast succeeds.

Ideally proper KTextEditor API should be introduced instead of filtering
events to make the functionality more reliable and future-proof. Yet, so
many events of the editor widget are filtered that new specialized APIs
are unlikely to completely replace the need to access the editor widget
any time soon.

So introduce two editor-widget member functions of KTextEditor::View to
at least prevent potential breakage when the editor widget is obtained
from a View and vice versa.

Install an event filter on View::editorWidget() rather than on
View::focusProxy() in KateTemplateHandler::setupEventHandler().

Remove 3 identical internal-linkage functions findViewInternal() in unit
tests that look for a KateViewInternal among the children of a View. Use
the new API View::editorWidget() in place of findViewInternal().

frameworks/ktexteditor#25